exercise pollitos-school functional methods  -create student  -get students  -create courses  -get courses
 The other methods requested in the exercise were completed and the use of logs was implemented.

changes that are considered based on the observations made

The changes discussed were applied during the course of Monday
