+++ src/main/resources/application.properties
@@ -2,5 +2,5 @@ spring.application.name=school-control
 
 school.score.limit=100
 
-spring.profiles.active=populated
+spring.profiles.active=default
